feat: add CommitmentsApiError variants for proof plan RPC refactor
The proof plan RPCs require knowledge of table schemas in storage.
Currently, they get this from the commitment metadata for tables in
storage, but we want them to get these schemas from the tables pallet
metadata instead. This path changes the sorts of error cases we run into
somewhat, which need to be reflected in the error enum and codes. This
adds the necessary variants, without reusing old codes to avoid a
breaking change.