feat: determine if an insert for a commitment exceeds end row limits
Proof-of-sql setups for different commitment schemes have different
limits regarding the end row of the table, aka the table's size. To
ensure we don't encroach upon this limit, we want to be able to reject
inserts that would exceed it in the runtime before attempting to commit.
This change adds some stateless logic in a new module in
pallet-commitments for determining if an insert exceeds some provided
limits, allowing for different limits per commitment scheme.