test: parameterize schemes in pallet-commitments table creation tests
The commitments pallet allows you to select a set of commitment schemes
when creating a table. However, all of the tests in the pallet simply
use all of the commitment schemes. It will be handy to be able to create
tests that only use some commitment schemes, especially when testing
situations where they reject differently (like scalar bounds, or table
size limits).