feat: add eth-ecdsa as a 4th runtime signature type
While we can already sign transactions with ECDSA, the semantics of
ECDSA signing are a bit different from ethereum. This change adds a
crate with a new signature type for ECDSA signatures that are a bit more
"ethereum compatible", as they expect payloads to be signed following
EIP-191. It also adds a new MultiSignature type whose encoding is
backwards compatible with sp_runtime's MultiSignature, just with an
extra EthEcdsa variant.