feat: add commitment-column-mapping crate
There aren't many utilities for editing commitments currently. We
haven't had to yet. This adds a crate for editing commitments following
some of rust's functional paradigms on a columnar basis, along with
OnChainTables.

This will soon be used to alter the way VarChar columns are committed to
as a pure runtime upgrade to be more EVM-compatible, despite the fact
that commitments are computed in the native interfaces.