feat: statelessly temporarily convert varchar columns to varbinary
From the api docs..

Older versions of the chain's native interface compute commitments to VarChar columns in ways
that aren't easy to verify in the EVM. On the other hand, a similar column, VarBinary, computes
commitments in a more EVM-compatible way.

These utilities are meant to be used to temporarily convert varchar columns to varbinary before
computing commitments in the native interface. That way, using a more EVM-commpatible VarChar
commitment strategy can be deployed as a runtime upgrade, and native interface improvements can
come in more gracefully.