fix: temporarily convert varchar columns to varbinary for commitments
Older versions of the chain's native interface compute commitments to
VarChar columns in ways that aren't easy to verify in the EVM. On the
other hand, a similar column, VarBinary, computes commitments in a more
EVM-compatible way.

This changes pallet-commitments to convert varchar columns to varbinary
before they are committed to in the native interface. Then they are
converted back afterwards. This makes it so the varchar commitments are
new computed in an EVM-compatible as a pure runtime upgrade. A new
native API can be added later to commit to varchar columns like this
more directly, but will take some time to ensure wide network adoption.