feat: add GenericOverCommitment type for Vecs
Most of the common core rust generic types can be used in a
higher-kinded fashion with the commitment scheme abstractions. Vecs are
a noticeable omission. To this end, this change adds VecType. This will
be important soon as we want to use the commitment scheme abstractions
over TableCommitment's internal Vec<C>.