feat: upgrade proof-of-sql to 0.121.1
This newer version of proof-of-sql comes with many new features and
breaking changes to the proof plans. Commitments have had breaking
changes too, but steps have been made to ensure that our commitment
computation on-chain remains backwards compatible. For now, this is the
version of proof-of-sql we'd like to stabilize on between the prover,
verifier contracts, and proof-of-sql SDK.