test: add proptest strategies for CommitmentScheme and -Flags
proptest allows us to test a huge variety of cases, with the aim of
demonstrating that certain "properties" of functions hold. In proptest,
Strategy is the modular, composable trait to define how types should be
generated for these cases.

One of the main types we'll want to be able to generate is commitments.
In particular, types like TableCommitmentsPerScheme. An important
building block for this is having strategies for CommitmentScheme and
CommitmentSchemeFlags. This change adds those strategies.