test: add binary for generating native backwards compatibility cases
We must never break the functionality of native interfaces, as it will
break any new node's ability to sync up from genesis. We currently don't
have any CI that verifies this backwards compatibility requirement. We
want some test cases that hard-code the native pass-by serialization of
their input/output. That way, if the function breaks, or if even the
serialization breaks, it will be caught in tests.

This adds a binary to the native crate -
generate-backwards-compatibility-cases. This generates hundreds of
hard-coded cases for the process_insert native interface.