feat: attest to claimed unstakes
Unstaking depends on attestors attesting to the amount that was
unstaked. Previously they were attesting to how much was locked in the
staking pallet, but this is not the actual amount that was claimed after
unstaking, it is the amount that could be unstaked. Furthermore, the
attestations to an unstake need to be unique, so signing these unstake
amounts multiple times over many blocks is incorrect.

This change makes it so attestors attest to the actual amount that was
claimed (using the new ClaimedUnstakes api), and that they sign it with
the block number that the claim occurred in, not the current block
number.