fix: burn tokens that are withdrawn immediately
Currently, we burn tokens after they have been completely unstaked,
meaning that the funds have made it to ethereum and indexers have picked
up the successful "unstaked" event. However, in the time between
claiming and burning, substrate accounts will have the funds freed up.
This leaves a window for them to potentially make transfers with those
funds, which we don't want them to do. This wasn't a problem before
because they had no way to sign transactions with their staking account,
but now they can with the EthEcdsa signature scheme. This makes it so we
burn the funds immediately after withdrawal instead of waiting.