fix: use domain to distinguish commitment and unstake attestations
We do not have complete confidence that attackers will not be able to
create specific tables whose commitments will be attested in a way that
could satisfy the staking contract checks. There are a few ways that
this could be addressed, this does so by adding an AttestationDomain to
the message for commitments (and for any other future use cases), while
keeping the unstake message bare.