feat: recalculate pallet-indexing weights
The pallet-indexing weights haven't been recalculated in some time. We
expect these to be worse than the last time they were calculated because
we now need to de/serialize commitments on either side of the native
interface for the varchar commitment workaround and to enforce table
maximum size limits. Inaccurate weights can enable DoS attacks, so it is
important that we try to reflect the truth here.