feat: make block number optional in attest_block extrinsic
Historically, the attestations storage uses a block number as a key, and
this block number is intended to be the block whose state is being
attested to. However, the design of the attestations pallet has been
changing to address some security concerns, and the definition of the
block numbers used in the signature, the attestation object, and now the
attestation key, have become more use-case-specific.

In particular, now we want attestors to use the block number of the
state that was attested in the commitments use case, but just the
current block number in the claimed unstake use case. Support for this
has been added here by making the block number parameter optional, and
using the current block number by default.