fix: attest in current block for claimed unstakes
With recent updates to the attest_block extrinsic, we can now have
attestations appear for the current block in the pallet-attestations
storage. This is important for the event forwarder to be able to easily
see the current claimed unstake attestations it needs to forward. This
change adjusts the attestors to use a new subxt codegen, the current
block for claimed unstakes, and the processed block for commitments.