fix: always use the attested block number in attestation struct
Currently, attestors create the claimed unstake attestation struct using
the block number that is being processed, not the claim block number.
This is incorrect, and prevents pallet-attestations from being able to
verify the signature. This change parameterizes some functions in the
attestor further in order to construct the correct attestation object.