refactor: generate attestation signatures in create_attestation
There is something that all attestation use cases follow for now. The
block number that is used in the message that is signed is the same
block number that is used in the attestation struct. This property will
likely persist because pallet-attestations checks for it. So, there is
less risk of parameterizing the wrong block number in the wrong function
if the create attestation function necessarily follows this rule.

This changes create_attestation to create the signature used in the
attestation block rather than have it passed in. This prevents invalid
input where the signature and partial message and block number don't
match. It also greatly simplifies the other functions in this crate as
they are updated to adjust to this new logic.