fix: retrieve correct attestation list to update in attest_block
The attest_block extrinsic needs to insert the attestation into the
block that is either paramterized or the current block. While it is
inserting the attestation into the correct storage key, it is not using
the correct initial collection to insert to, instead grabbing the block
number from the attestation struct.