fix: make table_commitment_util fallible
Some types that claim to have type guarantees actually do not because
they do not perform validation during serde deserialization.
TableCommitments are one of these, so the type guarantee that they have
negative ranges doesn't necessarily hold. To avoid making assumptions
about the safety of the rest of the chain, we should avoid expecting
these guarantees to hold for now. This adjusts the new
table_commitment_map function to error out instead of panic on these
assumptions.