perf: enable cpu-perf for all non-aarch64 targets
The current strategy of only enabling cpu-perf for x86_64 targets seems
to have disabled it for recent node versions and benchmark runs. This
significantly impacts performance, which can be seen in pallet-indexing
weights. This change is less stingy about enabling the flag, and seems
to have fixed its use in benchmarks and the node.