feat: Introduce transactions to update quorums for individual tables and a group of tables by namespace (#94)
# Rationale for this change
We want to introduce the option to update the quorum configuration for a
given table or set of tables. This PR introduces two new extrinsics for
`update_table_quorum` and `update_schema_quorum` that replace the
existing quorum entry and emit an event for each update. The event
emitted includes the previous value (if one was set), the new value, and
the table identifier.