fix: add missing benchmarks and weights for pallet-attestations
Some pallet-attestations extrinsics were missing benchmarks, and no
weights have ever been calculated for it. This is likely because there
was a bug in the existing benchmarks that had to do with the discrepancy
between AccountIds in the mock and benchmark runtimes. This change
addresses the bug, adds the missing benchmarks, and calculates the
weights.