feat: add frame weight template
The way we currently use WeightInfo types doesn't align well with the
default weight module template used by the benchmarking cli. This is
because the template pallets we use used a different weight template
which seems to be the canonical one, but not the default. As a result,
every time weights were calculated, the file would still need some
manual intervention to get it to match the interface the pallets expect.
Not only is this painful, but it would be very difficult to make weight
calculation automated.

This adds the canonical `frame-weight-template.hbs` which can be
supplied to the benchmark cli with `--template`. The weights modules
produced with this template need 0 modification.