fix: standardize docs of pallet-indexing
The rust-docs for pallet storage, calls, and events are available in the
runtime metadata. These docs are used by tools that convert this
metadata into documentation like subxt explorer. So, providing a decent
initial API reference can be as simple as improving the rustdocs of
pallets. This does so for pallet-indexing, unifying the language and
style.