feat: reorganize submissions-finding-quorum storage with SubmissionsV1
Currently, submissions storage is keyed from the data hash to the
accounts, rather than keyed from an account to data hash. This design
choice makes it possible for submitters to disagree with themselves,
and input an unbounded number of unique submissions for the same batch
id.

The solution for this is to change the map from being..
```
BatchId -> DataHash -> QuorumScope -> [AccountId]
```

to...
```
BatchId -> QuorumScope -> AccountId -> DataHash
```

This change adds a new piece of storage SubmissionsV1 to reflect that
second mapping, and switches the pallet-indexing logic over to using it.
No data migration is provided, since submissions are fairly disposable
in practice. A future PR will handle clearing out the old submissions
(and even pruning the new submissions).