feat: add VARIANT_COUNT constant to QuorumScope
no_std rust currently lacks the reflection functionality necessary to
implicitly count the number of variants on an enum. This number can be
useful in situations where you want to understand how large a collection
that is keyed by QuorumScopes can be, like in weight calculation or
usages of clear_prefix.

This change adds a manual VARIANT_COUNT const to QuorumScope for this
purpose. It is defined as a usize since, when this is replaced by
core::mem::variant_count, that will also return usize.