feat: add Heavy monad for composing functions that incur weight
Sometimes, we need to estimate the weight of an operation we're
performing in pallets. In particular, many pallet hooks expect weights
as a return type. Estimation is somewhat easy with the help of DbWeight,
but when many database operations are performed in complex scenarios,
it's easy to lose track.

This adds a Heavy monad, which allows us to easily make tiny functions
that incur weight and compose them for complex scenarios, with the
composed weight being the sum. This is used heavily in the upcoming
submissions storage bounding changes.