feat: derive Hash for TableIdentifier
As it is an identifier, TableIdentifier is often used as the unique key
in storage. In tests, it may be useful to use it as the key of a more
standard collection, but this requires that it implement the Hash trait.
This change derives Hash for TableIdentifier, but only in std scenarios.